+++ src/main/java/Ui.java
@@ -1,0 +1,8 @@
+public class Ui {
+
+    private static final String HORIZONTAL_LINE = "____________________________________________________________";
+
+    public static void print(String message) {
+        System.out.print(HORIZONTAL_LINE + "\n" + message + "\n" + HORIZONTAL_LINE + "\n");
+    }
+}
